backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = "Backend API"
authors = ["Dylan Rogowsky <dylan@oxd.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django-configurations = "^2.5.1"
gunicorn = "^23.0.0"
django = "^5.1.1"
django-structlog = "^8.1.0"
django-cors-headers = "^4.4.0"
psycopg = "^3.2.2"
openpyxl = "^3.1.5"
httpx = "^0.27.2"
django-filter = "^24.3"
attrs = "^24.2.0"
django-database-prefix = "^0.1.0"
psycopg2 = "^2.9.9"
unidecode = "^1.3.8"
django-storages = "^1.14.4"
boto3 = "^1.35.21"
setuptools = "^75.1.0"
django-json-widget = "^2.0.1"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
openai = "^1.46.0"
langchain-community = "^0.3.0"
scipy = "^1.14.1"
pydantic = "^2.9.2"
faiss-cpu = "^1.8.0.post1"

[tool.poetry.group.dev.dependencies]
pylint = "^3.2.7"
pylint-django = "^2.5.5"
black = "^24.8.0"
flake8 = "^7.1.1"
flake8-bugbear = "^24.8.19"
django-querycount = "^0.8.3"
drf-yasg = "^1.21.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
